# Aplicação/backend/static_docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PM AI MVP API",
    "description": "\n    ## PM AI MVP - Sistema de Gestão de Projetos com IA\n    \n    API completa para gestão de projetos com funcionalidades avançadas de IA.\n    \n    ### Funcionalidades Principais:\n    \n    * **Autenticação**: Sistema de autenticação OAuth com Google\n    * **Projetos**: Gestão completa de projetos e portfólios\n    * **Equipes**: Gestão de membros da equipe e funções\n    * **Clientes**: Gestão de clientes e níveis de comunicação\n    * **Riscos**: Análise e gestão de riscos com IA\n    * **Checklists**: Sistema de checklists para projetos\n    * **Action Items**: Gestão de itens de ação e tarefas\n    \n    ### Arquitetura:\n    \n    * **Services Layer**: Lógica de negócio separada dos controllers\n    * **Repository Pattern**: Camada de abstração para acesso a dados\n    * **Cache Redis**: Sistema de cache para performance\n    * **Processamento Assíncrono**: Tarefas em background com Celery\n    * **IA Integration**: Integração com Gemini API para análise de riscos\n    \n    ### Tecnologias:\n    \n    * **FastAPI**: Framework web moderno e rápido\n    * **SQLAlchemy**: ORM robusto com suporte assíncrono\n    * **PostgreSQL**: Banco de dados relacional\n    * **Redis**: Sistema de cache\n    * **Celery**: Processamento assíncrono\n    * **Pydantic**: Validação de dados\n    ",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/security/security/events": {
      "get": {
        "tags": [
          "security",
          "security"
        ],
        "summary": "Listar Eventos de Segurança",
        "description": "Lista eventos de segurança recentes com filtros opcionais.\n    \n    **Parâmetros:**\n    - hours: Número de horas para buscar eventos (padrão: 24)\n    - severity: Filtrar por nível de severidade (low, medium, high, critical)\n    - threat_type: Filtrar por tipo de ameaça\n    - limit: Número máximo de eventos (padrão: 100)",
        "operationId": "get_security_events_api_v1_security_security_events_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "hours",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 168,
              "minimum": 1,
              "description": "Hours to look back",
              "default": 24,
              "title": "Hours"
            },
            "description": "Hours to look back"
          },
          {
            "name": "severity",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by severity level",
              "title": "Severity"
            },
            "description": "Filter by severity level"
          },
          {
            "name": "threat_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by threat type",
              "title": "Threat Type"
            },
            "description": "Filter by threat type"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "description": "Maximum number of events",
              "default": 100,
              "title": "Limit"
            },
            "description": "Maximum number of events"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  },
                  "title": "Response Get Security Events Api V1 Security Security Events Get"
                }
              }
            }
          },
          "404": {
            "description": "Recurso não encontrado"
          },
          "400": {
            "description": "Dados inválidos"
          },
          "401": {
            "description": "Não autorizado"
          },
          "403": {
            "description": "Sem permissão"
          },
          "500": {
            "description": "Erro interno do servidor"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/security/security/summary": {
      "get": {
        "tags": [
          "security",
          "security"
        ],
        "summary": "Resumo de Segurança",
        "description": "Obtém resumo estatístico de eventos de segurança.\n    \n    **Retorna:**\n    - Contadores de eventos por tipo e severidade\n    - IPs bloqueados\n    - Configurações de rate limiting\n    - Estatísticas de 24h e 7 dias",
        "operationId": "get_security_summary_api_v1_security_security_summary_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Get Security Summary Api V1 Security Security Summary Get"
                }
              }
            }
          },
          "404": {
            "description": "Recurso não encontrado"
          },
          "400": {
            "description": "Dados inválidos"
          },
          "401": {
            "description": "Não autorizado"
          },
          "403": {
            "description": "Sem permissão"
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/security/security/blocked-ips": {
      "get": {
        "tags": [
          "security",
          "security"
        ],
        "summary": "Listar IPs Bloqueados",
        "description": "Lista todos os IPs atualmente bloqueados por ameaças de segurança.",
        "operationId": "get_blocked_ips_api_v1_security_security_blocked_ips_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "type": "string"
                  },
                  "type": "array",
                  "title": "Response Get Blocked Ips Api V1 Security Security Blocked Ips Get"
                }
              }
            }
          },
          "404": {
            "description": "Recurso não encontrado"
          },
          "400": {
            "description": "Dados inválidos"
          },
          "401": {
            "description": "Não autorizado"
          },
          "403": {
            "description": "Sem permissão"
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/security/security/blocked-ips/{ip_address}": {
      "post": {
        "tags": [
          "security",
          "security"
        ],
        "summary": "Bloquear IP",
        "description": "Bloqueia um IP específico por ameaças de segurança.\n    \n    **Parâmetros:**\n    - ip_address: Endereço IP para bloquear",
        "operationId": "block_ip_api_v1_security_security_blocked_ips__ip_address__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "ip_address",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ip Address"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Block Ip Api V1 Security Security Blocked Ips  Ip Address  Post"
                }
              }
            }
          },
          "404": {
            "description": "Recurso não encontrado"
          },
          "400": {
            "description": "Dados inválidos"
          },
          "401": {
            "description": "Não autorizado"
          },
          "403": {
            "description": "Sem permissão"
          },
          "500": {
            "description": "Erro interno do servidor"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "security",
          "security"
        ],
        "summary": "Desbloquear IP",
        "description": "Remove o bloqueio de um IP específico.\n    \n    **Parâmetros:**\n    - ip_address: Endereço IP para desbloquear",
        "operationId": "unblock_ip_api_v1_security_security_blocked_ips__ip_address__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "ip_address",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ip Address"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Unblock Ip Api V1 Security Security Blocked Ips  Ip Address  Delete"
                }
              }
            }
          },
          "404": {
            "description": "Recurso não encontrado"
          },
          "400": {
            "description": "Dados inválidos"
          },
          "401": {
            "description": "Não autorizado"
          },
          "403": {
            "description": "Sem permissão"
          },
          "500": {
            "description": "Erro interno do servidor"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/security/security/rate-limits": {
      "get": {
        "tags": [
          "security",
          "security"
        ],
        "summary": "Status dos Rate Limits",
        "description": "Obtém status atual dos rate limits para diferentes tipos de endpoint.",
        "operationId": "get_rate_limits_status_api_v1_security_security_rate_limits_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Get Rate Limits Status Api V1 Security Security Rate Limits Get"
                }
              }
            }
          },
          "404": {
            "description": "Recurso não encontrado"
          },
          "400": {
            "description": "Dados inválidos"
          },
          "401": {
            "description": "Não autorizado"
          },
          "403": {
            "description": "Sem permissão"
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/security/security/password-strength": {
      "get": {
        "tags": [
          "security",
          "security"
        ],
        "summary": "Validar Força da Senha",
        "description": "Valida a força de uma senha sem armazená-la.\n    \n    **Parâmetros:**\n    - password: Senha para validar",
        "operationId": "validate_password_strength_api_v1_security_security_password_strength_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "password",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Password to validate",
              "title": "Password"
            },
            "description": "Password to validate"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Validate Password Strength Api V1 Security Security Password Strength Get"
                }
              }
            }
          },
          "404": {
            "description": "Recurso não encontrado"
          },
          "400": {
            "description": "Dados inválidos"
          },
          "401": {
            "description": "Não autorizado"
          },
          "403": {
            "description": "Sem permissão"
          },
          "500": {
            "description": "Erro interno do servidor"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/security/security/csrf-token": {
      "post": {
        "tags": [
          "security",
          "security"
        ],
        "summary": "Gerar Token CSRF",
        "description": "Gera um token CSRF para proteção contra ataques CSRF.",
        "operationId": "generate_csrf_token_api_v1_security_security_csrf_token_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Generate Csrf Token Api V1 Security Security Csrf Token Post"
                }
              }
            }
          },
          "404": {
            "description": "Recurso não encontrado"
          },
          "400": {
            "description": "Dados inválidos"
          },
          "401": {
            "description": "Não autorizado"
          },
          "403": {
            "description": "Sem permissão"
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/security/security/login-attempts/{identifier}": {
      "get": {
        "tags": [
          "security",
          "security"
        ],
        "summary": "Status de Tentativas de Login",
        "description": "Verifica o status de tentativas de login para um usuário ou IP.\n    \n    **Parâmetros:**\n    - identifier: ID do usuário ou endereço IP",
        "operationId": "get_login_attempts_status_api_v1_security_security_login_attempts__identifier__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Identifier"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Get Login Attempts Status Api V1 Security Security Login Attempts  Identifier  Get"
                }
              }
            }
          },
          "404": {
            "description": "Recurso não encontrado"
          },
          "400": {
            "description": "Dados inválidos"
          },
          "401": {
            "description": "Não autorizado"
          },
          "403": {
            "description": "Sem permissão"
          },
          "500": {
            "description": "Erro interno do servidor"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}